requirements.txt
Django>=3.1
inflection>=0.5.0
ruamel.yaml>=0.16.10
addict>=2.2.1
bump2version>=1.0.0
black>=19.10b0
pre-commit>=2.6.0
flake8>=3.8.3
